test: improve type definition and tests
