--- packages/preset-dumi/src/transformer/fixtures/remark-api/Hello/World.tsx
@@ -1,1 +1,0 @@
-export { default } from '.';
