--- packages/preset-dumi/src/transformer/fixtures/remark-api/Hello/index.md
@@ -1,5 +1,0 @@
----
-componentName: Helloooo
----
-
-<API />
