--- packages/preset-dumi/src/transformer/fixtures/remark-api/auto-detect/index.md
@@ -1,1 +1,0 @@
-<API />
