--- packages/preset-dumi/src/transformer/fixtures/remark-api/custom-exports.md
@@ -1,1 +1,0 @@
-<API src="./Hello/index.tsx" exports='["default"]'></API>
