--- packages/preset-dumi/src/transformer/fixtures/remark-api/hide-title.md
@@ -1,1 +1,0 @@
-<API src="@/Hello/index.tsx" hideTitle/>
