--- packages/preset-dumi/src/transformer/fixtures/remark-api/packages/pkgA/src/index.tsx
@@ -1,7 +1,0 @@
-import type React from 'react';
-
-const Hello: React.FC<{ className?: string }> = () => null;
-
-export const World = Hello;
-
-export default Hello;
