+++ packages/preset-dumi/src/transformer/fixtures/remark-component/DumiApi.md
@@ -1,0 +1,1 @@
+<DumiApi />
